make shared module android only
