+++ shared/build.gradle.kts
@@ -1,8 +1,7 @@
-@Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.compose)
-    alias(libs.plugins.kotlin.multiplatform)
+    alias(libs.plugins.kotlin.android)
     alias(libs.plugins.kotlin.serialization)
     alias(libs.plugins.aboutlibraries)
 }
@@ -15,35 +14,22 @@ android {
         minSdk = 21
     }
 
-    buildFeatures {
-        buildConfig = true
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
     }
 }
 
-kotlin {
-    androidTarget()
-    jvm("desktop")
+dependencies {
+    api(compose.runtime)
+    api(libs.bundles.kotlinx)
 
-    jvmToolchain(17)
+    api(libs.aboutlibraries.core)
+    api(libs.apollo.runtime)
+    api(libs.apollo.normalized.cache)
+    api(libs.koin.core)
+    api(libs.multiplatform.settings)
 
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(compose.runtime)
-                api(libs.bundles.kotlinx)
-
-                api(libs.aboutlibraries.core)
-                api(libs.apollo.runtime)
-                api(libs.apollo.normalized.cache)
-                api(libs.koin.core)
-                api(libs.multiplatform.settings)
-            }
-        }
-        val androidMain by getting {
-            dependencies {
-                implementation(libs.androidx.browser)
-                implementation(libs.androidx.core.ktx)
-            }
-        }
-    }
+    implementation(libs.androidx.browser)
+    implementation(libs.androidx.core.ktx)
 }
