--- shared/src/androidMain/kotlin/com/materiiapps/gloom/utils/SettingsProvider.kt
@@ -1,16 +1,0 @@
-package com.materiiapps.gloom.utils
-
-import android.content.Context
-import com.russhwolf.settings.Settings
-import com.russhwolf.settings.SharedPreferencesSettings
-
-actual class SettingsProvider(
-    private val context: Context,
-    private val name: String
-) {
-
-    actual fun createSettings(): Settings = SharedPreferencesSettings(
-        context.getSharedPreferences(name, Context.MODE_PRIVATE)
-    )
-
-}
