--- shared/src/commonMain/kotlin/com/materiiapps/gloom/di/modules/ManagerModule.kt
@@ -1,5 +1,0 @@
-package com.materiiapps.gloom.di.modules
-
-import org.koin.core.module.Module
-
-expect fun managerModule(): Module
