--- shared/src/commonMain/kotlin/com/materiiapps/gloom/domain/manager/DownloadManager.kt
@@ -1,5 +1,0 @@
-package com.materiiapps.gloom.domain.manager
-
-expect class DownloadManager {
-    fun download(url: String, block: (String) -> Unit = {})
-}
