--- shared/src/commonMain/kotlin/com/materiiapps/gloom/domain/manager/ShareManager.kt
@@ -1,7 +1,0 @@
-package com.materiiapps.gloom.domain.manager
-
-expect class ShareManager {
-
-    fun shareText(text: String)
-
-}
