--- shared/src/commonMain/kotlin/com/materiiapps/gloom/utils/Logger.kt
@@ -1,12 +1,0 @@
-package com.materiiapps.gloom.utils
-
-expect class Logger() {
-
-    fun error(tag: String, message: String?, throwable: Throwable?)
-
-    fun warn(tag: String, message: String, throwable: Throwable?)
-
-    fun info(tag: String, message: String)
-
-    fun debug(tag: String, message: String)
-}
