--- shared/src/commonMain/kotlin/com/materiiapps/gloom/utils/PlatformUtils.kt
@@ -1,11 +1,0 @@
-package com.materiiapps.gloom.utils
-
-import java.io.File
-
-expect val supportsMonet: Boolean
-expect val isDebug: Boolean
-
-expect val Features: List<Feature>
-expect val GloomPath: File
-
-
