--- shared/src/commonMain/kotlin/com/materiiapps/gloom/utils/SettingsProvider.kt
@@ -1,9 +1,0 @@
-package com.materiiapps.gloom.utils
-
-import com.russhwolf.settings.Settings
-
-expect class SettingsProvider {
-
-    fun createSettings(): Settings
-
-}
