--- shared/src/desktopMain/kotlin/com/materiiapps/gloom/domain/manager/DownloadManager.kt
@@ -1,9 +1,0 @@
-package com.materiiapps.gloom.domain.manager
-
-actual class DownloadManager {
-
-    actual fun download(url: String, block: (String) -> Unit) {
-        // TODO: Figure out best method for implementation
-    }
-
-}
