--- shared/src/desktopMain/kotlin/com/materiiapps/gloom/domain/manager/LibraryManager.kt
@@ -1,9 +1,0 @@
-package com.materiiapps.gloom.domain.manager
-
-import com.mikepenz.aboutlibraries.Libs
-
-actual class LibraryManager {
-
-    actual val libs = Libs.Builder().build()
-
-}
