--- shared/src/desktopMain/kotlin/com/materiiapps/gloom/domain/manager/ShareManager.kt
@@ -1,9 +1,0 @@
-package com.materiiapps.gloom.domain.manager
-
-actual class ShareManager {
-
-    actual fun shareText(text: String) {
-        // TODO: Implement
-    }
-
-}
