+++ shared/src/main/kotlin/com/materiiapps/gloom/domain/manager/DialogManager.kt
