+++ shared/src/main/kotlin/com/materiiapps/gloom/domain/manager/base/BasePreferenceManager.kt
